Moving stuff around and starting on new data objects
