+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/emails/ActivityFeedEmailSender.scala
@@ -20,6 +20,7 @@ import com.keepit.common.performance._
 import com.keepit.curator.LibraryQualityHelper
 import com.keepit.eliza.ElizaServiceClient
 import com.keepit.model._
+import com.keepit.model.keep.KeepInfo
 import com.keepit.social.BasicUser
 import play.twirl.api.Html
