+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/LibraryController.scala
@@ -24,6 +24,7 @@ import com.keepit.heimdal.HeimdalContextBuilderFactory
 import com.keepit.inject.FortyTwoConfig
 import com.keepit.model.ExternalLibrarySpace.{ ExternalOrganizationSpace, ExternalUserSpace }
 import com.keepit.model._
+import com.keepit.model.keep.KeepInfo
 import com.keepit.shoebox.controllers.LibraryAccessActions
 import com.keepit.social.BasicUser
 import org.joda.time.DateTime
