+++ kifi-backend/modules/shoebox/app/com/keepit/model/FullLibraryInfo.scala
@@ -4,6 +4,7 @@ import com.keepit.common.crypto.{ PublicId, PublicIdConfiguration }
 import com.keepit.common.json
 import com.keepit.common.mail.BasicContact
 import com.keepit.discussion.Message
+import com.keepit.model.keep.KeepInfo
 import com.keepit.slack.LibrarySlackInfo
 import com.keepit.slack.models.{ SlackIntegrationStatus, SlackChannelToLibrary, LibraryToSlackChannel, SlackChannelName }
 import com.keepit.social.BasicUser
